requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
attrs==25.1.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector==2.2.9
mysqlclient==2.2.7
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
